Refactored server statistics to just be 'stats', changed responses to just send back 'stats' for each, and removed log
