+++ server/db/init.js
@@ -3,7 +3,7 @@
 const teams = require('./teams')
 const submissions = require('./submissions')
 const games = require('./games')
-const stats = require('./statistics')
+const stats = require('./stats')
 
 module.exports = {
   teams,
