+++ server/routers/init.js
@@ -13,7 +13,7 @@ const submissions = require('./submissions')
 const games = require('./games')
 const login = require('./login')
 const medium = require('./medium')
-const stats = require('./statistics')
+const stats = require('./stats')
 const admin = require('./admin')
 
 module.exports = {
